EKF use ecl_float_t type (double precision floating point where available)

+++ src/lib/mathlib/math/Limits.hpp
@@ -102,7 +102,7 @@ inline bool isZero(float val)
 /** Safe way to check if double is zero */
 inline bool isZero(double val)
 {
-	return fabs(val - 0.0) < DBL_EPSILON;
+	return fabs(val - (double)0.0) < DBL_EPSILON;
 }
 
 }
